requirements.txt
google-api-core==1.31.0
google-auth==1.33.0
google-cloud==0.34.0
google-cloud-core==1.7.1
google-cloud-pubsub==2.6.1
google-cloud-storage==1.41.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.38.1
mccabe==0.6.1
pandas==1.3.0
python-dateutil==2.8.2
python-decouple==3.4
requests==2.26.0